reshape api vehicle occupancies and modify some module
