+++ internal/utils/utils_test.go
@@ -1,12 +1,66 @@
 package utils
 
 import (
+	"flag"
+	"fmt"
+	"net/url"
+	"os"
 	"testing"
 	"time"
 
+	"github.com/ory/dockertest"
+	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+const oneline = "1;87A;Mions Bourdelle;11 min;E;2018-09-17 20:28:00;35998;87A-022AM:5:2:12\r\n"
+
+var defaultTimeout time.Duration = time.Second * 10
+
+var sftpPort string
+var fixtureDir string
+
+func TestMain(m *testing.M) {
+
+	fixtureDir = os.Getenv("FIXTUREDIR")
+	if fixtureDir == "" {
+		panic("$FIXTUREDIR isn't set")
+	}
+
+	flag.Parse() //required to get Short() from testing
+	if testing.Short() {
+		logrus.Warn("skipping test Docker in short mode.")
+		os.Exit(m.Run())
+	}
+
+	pool, err := dockertest.NewPool("")
+	if err != nil {
+		logrus.Fatalf("Could not connect to docker: %s", err)
+	}
+	opts := dockertest.RunOptions{
+		Repository: "atmoz/sftp",
+		Tag:        "alpine",
+		Cmd:        []string{"forseti:pass"},
+		Mounts:     []string{fmt.Sprintf("%s:/home/forseti", fixtureDir)},
+	}
+	resource, err := pool.RunWithOptions(&opts)
+	if err != nil {
+		logrus.Fatalf("Could not start resource: %s", err)
+	}
+	//lets wait a bit for the docker to start :(
+	time.Sleep(3 * time.Second)
+	sftpPort = resource.GetPort("22/tcp")
+	//Run tests
+	code := m.Run()
+
+	// You can't defer this because os.Exit doesn't care for defer
+	if err := pool.Purge(resource); err != nil {
+		logrus.Fatalf("Could not purge resource: %s", err)
+	}
+
+	os.Exit(code)
+}
 
 func TestStringToInt(t *testing.T) {
 
@@ -44,4 +98,81 @@ func TestCalculateOccupancy(t *testing.T) {
 	assert.Equal(result, 0)
 	result = CalculateOccupancy(50)
 	assert.Equal(result, 50)
+}
+
+func TestGetSFTPFileTimeout(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping test Docker in short mode.")
+	}
+	uri, err := url.Parse(fmt.Sprintf("sftp://forseti:pass@localhost:%s/oneline.txt", sftpPort))
+	require.Nil(t, err)
+	_, err = GetFileWithSftp(*uri, time.Microsecond)
+	require.NotNil(t, err)
+}
+func TestGetSFTPFile(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping test Docker in short mode.")
+	}
+	uri, err := url.Parse(fmt.Sprintf("sftp://forseti:pass@localhost:%s/oneline.txt", sftpPort))
+	require.Nil(t, err)
+	reader, err := GetFileWithSftp(*uri, defaultTimeout)
+	require.Nil(t, err)
+	b := make([]byte, 100)
+	len, err := reader.Read(b)
+	assert.Nil(t, err)
+	assert.Equal(t, oneline, string(b[0:len]))
+
+	reader, err = GetFile(*uri, defaultTimeout)
+	assert.Nil(t, err)
+	len, err = reader.Read(b)
+	assert.Nil(t, err)
+	assert.Equal(t, oneline, string(b[0:len]))
+}
+
+func TestGetSFTPFS(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping test Docker in short mode.")
+	}
+	uri, err := url.Parse(fmt.Sprintf("file://%s/oneline.txt", fixtureDir))
+	require.Nil(t, err)
+	reader, err := GetFileWithFS(*uri)
+	require.Nil(t, err)
+	b := make([]byte, 100)
+	len, err := reader.Read(b)
+	assert.Nil(t, err)
+	assert.Equal(t, oneline, string(b[0:len]))
+
+	reader, err = GetFile(*uri, defaultTimeout)
+	assert.Nil(t, err)
+	len, err = reader.Read(b)
+	assert.Nil(t, err)
+	assert.Equal(t, oneline, string(b[0:len]))
+
+}
+
+func TestGetSFTPFileError(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping test Docker in short mode.")
+	}
+	uri, err := url.Parse(fmt.Sprintf("sftp://forseti:wrongpass@localhost:%s/oneline.txt", sftpPort))
+	require.Nil(t, err)
+	_, err = GetFileWithSftp(*uri, defaultTimeout)
+	require.Error(t, err)
+
+	_, err = GetFile(*uri, defaultTimeout)
+	require.Error(t, err)
+
+	uri, err = url.Parse(fmt.Sprintf("sftp://monuser:pass@localhost:%s/oneline.txt", sftpPort))
+	require.Nil(t, err)
+	_, err = GetFileWithSftp(*uri, defaultTimeout)
+	require.Error(t, err)
+	_, err = GetFile(*uri, defaultTimeout)
+	require.Error(t, err)
+
+	uri, err = url.Parse(fmt.Sprintf("sftp://forseti:pass@localhost:%s/not.txt", sftpPort))
+	require.Nil(t, err)
+	_, err = GetFileWithSftp(*uri, defaultTimeout)
+	require.Error(t, err)
+	_, err = GetFile(*uri, defaultTimeout)
+	require.Error(t, err)
 }
